feat: integrate adaptive storefront runtime into VipBike landing + ext
Apply AI assistant changes to 14 files.
Ref: feat: integrate adaptive storefront runtime into VipBike landing + extract franchise features...
